ASR/requirements.txt
pvporcupine 
sounddevice 
webrtcvad 
numpy
python-dotenv
fastapi
uvicorn
httpx